refactor: 유효성 검사 prefix validate -> check 변경
